Rename subcommand package to skeleton
