+++ docs/project/changelog.md
@@ -95,8 +95,8 @@ substitutions:
 - {{ Fix }} Shared libraries with version suffix are now handled correctly.
   {pr}`3154`
 
-- Added a new CLI command `pyodide package`.
-  `pyodide-build mkpkg` will be replaced by `pyodide package`.
+- Added a new CLI command `pyodide sekeleton` which creates a package build recipe.
+  `pyodide-build mkpkg` will be replaced by `pyodide sekeleton pypi`.
   {pr}`3175`
 
 ### Build System / Package Loading
