+++ pyodide-build/setup.cfg
@@ -39,7 +39,7 @@ console_scripts =
 pyodide.cli =
     build = pyodide_build.cli.build_oot:main
     venv = pyodide_build.cli.venv:main
-    package = pyodide_build.cli.new_package:app
+    skeleton = pyodide_build.cli.skeleton:app
 
 [options.extras_require]
 test =
